PPO kinda works?
PPO takes 10 minutes for me to train on GPU and 8 after round 0. The actor and critic loss in round 0 is really high but get's normalized in round 2. I am not sure if it learns. But I'll leave it training for a bit, but I feel something is wrong. Because a 500 actor loss at the start that gets normalized to near 0 seems really strange to me